+++ compiler/bin-js_of_ocaml/print_runtime.ml
@@ -1,6 +1,6 @@
 (* Js_of_ocaml compiler
  * http://www.ocsigen.org/js_of_ocaml/
- * Copyright (C) 2014 Hugo Heuzard
+ * Copyright (C) 2020 Hugo Heuzard
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,35 +17,22 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *)
 
+open! Js_of_ocaml_compiler.Stdlib
 open Js_of_ocaml_compiler
 
-type t =
-  { common : Jsoo_cmdline.Arg.t
-  ; (* compile option *)
-    profile : Driver.profile option
-  ; source_map : (string option * Source_map.t) option
-  ; runtime_files : string list
-  ; no_runtime : bool
-  ; runtime_only : bool
-  ; output_file : [ `Name of string | `Stdout ] * bool
-  ; input_file : string option
-  ; params : (string * string) list
-  ; static_env : (string * string) list
-  ; wrap_with_fun : string option
-  ; (* toplevel *)
-    dynlink : bool
-  ; linkall : bool
-  ; toplevel : bool
-  ; export_file : string option
-  ; nocmis : bool
-  ; (* filesystem *)
-    include_dir : string list
-  ; fs_files : string list
-  ; fs_output : string option
-  ; fs_external : bool
-  ; keep_unit_names : bool
-  }
+let f () =
+  let output = stdout in
+  let new_line () = output_string output "\n" in
+  List.iter Jsoo_runtime.runtime ~f:(fun x ->
+      output_string output (Printf.sprintf "//# 1 %S" (Builtins.File.name x));
+      new_line ();
+      output_string output (Builtins.File.content x);
+      new_line ())
 
-val options : t Cmdliner.Term.t
+let info =
+  Info.make
+    ~name:"print-standard-runtime"
+    ~doc:"Print standard runtime to stdout"
+    ~description:"js_of_ocaml-print-standard-runtime dump the standard runtime to stdout."
 
-val options_runtime_only : t Cmdliner.Term.t
+let command = Cmdliner.Term.(pure f $ pure ()), info