+++ compiler/lib/builtins.ml
@@ -1,6 +1,6 @@
 (* Js_of_ocaml compiler
  * http://www.ocsigen.org/js_of_ocaml/
- * Copyright (C) 2013 Hugo Heuzard
+ * Copyright (C) 2020 Hugo Heuzard
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -16,29 +16,32 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *)
+open Stdlib
 
-type t =
-  { src : string option
-  ; name : string option
-  ; col : int
-  ; line : int
-  ; idx : int
-  ; fol : bool option
-  }
+module File = struct
+  (* TODO: Storing more information here (e.g. primitives with arity &
+     purity) would allow us to not parse the file if we're not
+     linking. *)
+  type t =
+    { name : string
+    ; content : string
+    }
 
-val zero : t
+  let name t = t.name
 
-type lineinfo
+  let content t = t.content
+end
 
-val make_lineinfo_from_file : string -> lineinfo
+let tbl = Hashtbl.create 17
 
-val make_lineinfo_from_string :
-  ?name:string -> ?src:string -> ?offset:t -> string -> lineinfo
+let register ~name ~content =
+  let name = "+" ^ name in
+  let t = { File.name; content } in
+  if Hashtbl.mem tbl name
+  then warn "The builtin runtime file %S was registered multiple time" name;
+  Hashtbl.add tbl name t;
+  t
 
-val make_lineinfo_from_channel : in_channel -> lineinfo * string
+let find name = try Some (Hashtbl.find tbl name) with Not_found -> None
 
-val relative_path : lineinfo -> string -> string option
-
-val t_of_lexbuf : lineinfo -> Lexing.lexbuf -> t
-
-val t_of_position : src:string option -> Lexing.position -> t
+let all () = Hashtbl.fold (fun _ v acc -> v :: acc) tbl []