+++ compiler/lib/builtins.mli
@@ -1,6 +1,6 @@
 (* Js_of_ocaml compiler
  * http://www.ocsigen.org/js_of_ocaml/
- * Copyright (C) 2013 Hugo Heuzard
+ * Copyright (C) 2020 Hugo Heuzard
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,28 +17,16 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *)
 
-type t =
-  { src : string option
-  ; name : string option
-  ; col : int
-  ; line : int
-  ; idx : int
-  ; fol : bool option
-  }
+module File : sig
+  type t
 
-val zero : t
+  val name : t -> string
 
-type lineinfo
+  val content : t -> string
+end
 
-val make_lineinfo_from_file : string -> lineinfo
+val find : string -> File.t option
 
-val make_lineinfo_from_string :
-  ?name:string -> ?src:string -> ?offset:t -> string -> lineinfo
+val all : unit -> File.t list
 
-val make_lineinfo_from_channel : in_channel -> lineinfo * string
-
-val relative_path : lineinfo -> string -> string option
-
-val t_of_lexbuf : lineinfo -> Lexing.lexbuf -> t
-
-val t_of_position : src:string option -> Lexing.position -> t
+val register : name:string -> content:string -> File.t